+++ qunicorn_core/api/deployment_api/deployment_view.py
@@ -20,8 +20,9 @@ from http import HTTPStatus
 from flask.views import MethodView
 
 from .root import DEPLOYMENT_API
+from ..api_models import JobResponseDtoSchema
 from ..api_models.deployment_dtos import DeploymentDtoSchema, DeploymentRequestDtoSchema, DeploymentRequestDto
-from ...core.jobmanager import deployment_service
+from ...core.jobmanager import deployment_service, jobmanager_service
 from ...util import logging
 
 
@@ -67,3 +68,20 @@ class DeploymentDetailView(MethodView):
         logging.info("Request: update deployment by id")
         deployment_dto: DeploymentRequestDto = DeploymentRequestDto.from_dict(body)
         return deployment_service.update_deployment(deployment_dto, deployment_id)
+
+
+@DEPLOYMENT_API.route("/<string:deployment_id>/jobs")
+class JobsByDeploymentView(MethodView):
+    """API endpoint for jobs of a specific deployment."""
+
+    @DEPLOYMENT_API.response(HTTPStatus.OK, JobResponseDtoSchema(many=True))
+    def get(self, deployment_id: str):
+        """Get the details of all jobs with a specific deployment id."""
+        logging.info("Request: get jobs with deployment id")
+        return jobmanager_service.get_jobs_by_deployment_id(deployment_id)
+
+    @DEPLOYMENT_API.response(HTTPStatus.OK, JobResponseDtoSchema(many=True))
+    def delete(self, deployment_id: str):
+        """Delete all jobs with a specific deployment id."""
+        logging.info("Request: delete jobs with deployment id")
+        return jobmanager_service.delete_jobs_by_deployment_id(deployment_id)
